move in file.ts
